Start work on data driven sidebar
