--- packages/app/src/domain/layout/logic/sidebar.ts
@@ -1,12 +1,0 @@
-type Category = {
-  title: string;
-  items: Item[];
-};
-
-type Item = {
-  title: string;
-  icon: React.ReactElement;
-  href: string;
-};
-
-export type SidebarData = Category | Item[];
